requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.1.1
Automat==20.2.0
Babel==2.9.0
boto3==1.14.57
botocore==1.17.63
certifi==2020.12.5
cffi==1.14.4
channels==2.4.0
channels-redis==2.4.2
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.4
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-places==4.0.0
Django==3.0.7
django-allauth==0.42.0
django-braces==1.14.0
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-filter==2.4.0
django-import-export==2.0.2
django-js-asset==1.2.2
django-markdown==0.8.4
django-model-utils==4.0.0
django-modeltranslation==0.15
django-paypal==1.0.0
django-phonenumber-field==4.0.0
django-redis==4.12.1
django-rest-auth==0.9.5
django-storages==1.10
django-taggit==1.3.0
djangorestframework==3.12.2
docutils==0.15.2
et-xmlfile==1.0.1
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
jdcal==1.4.1
jmespath==0.10.0
Markdown==3.3.3
MarkupPy==1.14
msgpack==0.6.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
Pillow==7.1.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.4.1
redis==3.0.0
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
Twisted==20.3.0
txaio==20.12.1
urllib3==1.25.11
websockets==7.0
whitenoise==5.1.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.2.0
